Refactor core app structure and remove obsolete code
