+++ .github/workflows/release.yaml
@@ -153,11 +153,11 @@ jobs:
           path: |
             src-tauri/binaries/ariso-stt-aarch64-apple-darwin
             src-tauri/binaries/mlx-swift_Cmlx.bundle
-          key: ariso-stt-aarch64-${{ steps.toolchain.outputs.fp }}-${{ hashFiles('src-tauri/ariso-stt/Package.resolved', 'src-tauri/ariso-stt/Sources/**') }}
+          key: ariso-stt-aarch64-${{ steps.toolchain.outputs.fp }}-${{ hashFiles('src-tauri/ariso-stt-mac/Package.resolved', 'src-tauri/ariso-stt-mac/Sources/**') }}
 
       - name: Build ariso-stt sidecar
         if: steps.sidecar-cache.outputs.cache-hit != 'true'
-        working-directory: src-tauri/ariso-stt
+        working-directory: src-tauri/ariso-stt-mac
         run: |
           mkdir -p ../binaries
           xcodebuild build -scheme ariso-stt -configuration Release \
@@ -262,7 +262,7 @@ jobs:
         run: npm ci --prefer-offline
 
       - name: Test Windows ariso-stt sidecar
-        run: cargo test --manifest-path src-tauri/ariso-stt-cross/Cargo.toml --locked
+        run: cargo test --manifest-path src-tauri/ariso-stt-windows/Cargo.toml --locked
 
       - name: Build Windows ariso-stt sidecar
         shell: pwsh
