+++ .gitignore
@@ -3,12 +3,11 @@ node_modules/
 output/
 src-tauri/target/
 src-tauri/binaries/
-src-tauri/ariso-stt/.build/
-src-tauri/ariso-stt/.xcode/
-src-tauri/ariso-stt/.xcode-metallib/
-src-tauri/ariso-stt/.swiftpm/
-src-tauri/ariso-stt-cross/target/
-.codex-artifacts/
+src-tauri/ariso-stt-mac/.build/
+src-tauri/ariso-stt-mac/.xcode/
+src-tauri/ariso-stt-mac/.xcode-metallib/
+src-tauri/ariso-stt-mac/.swiftpm/
+src-tauri/ariso-stt-windows/target/
 .superpowers/
 docs/superpowers/plans/
 .DS_Store
