+++ docs/superpowers/specs/2026-06-03-local-meeting-notes-design.md
@@ -51,7 +51,7 @@ returns data on stdout and Rust persists it.
 
 ## Components
 
-### Swift sidecar (`src-tauri/ariso-stt`)
+### Swift sidecar (`src-tauri/ariso-stt-mac`)
 
 **Dependency:** add `https://github.com/ml-explore/mlx-swift-lm` to `Package.swift`
 (products `MLXLLM`, `MLXLMCommon`, `MLXHuggingFace`), which transitively pulls in
