+++ scripts/build-windows-sidecar.ps1
@@ -14,14 +14,14 @@ try {
   Import-WindowsBuildEnvironment
 
   cargo "+$Toolchain" build `
-    --manifest-path src-tauri/ariso-stt-cross/Cargo.toml `
+    --manifest-path src-tauri/ariso-stt-windows/Cargo.toml `
     --release `
     --locked `
     --target $Target
 
   New-Item -ItemType Directory -Force src-tauri/binaries | Out-Null
   Copy-Item `
-    "src-tauri/ariso-stt-cross/target/$Target/release/ariso-stt.exe" `
+    "src-tauri/ariso-stt-windows/target/$Target/release/ariso-stt.exe" `
     "src-tauri/binaries/ariso-stt-$Target.exe" `
     -Force
 } finally {
