+++ src-tauri/ariso-stt-mac/Package.resolved
