+++ src-tauri/ariso-stt-mac/Package.swift
@@ -2,7 +2,7 @@
 import PackageDescription
 
 let package = Package(
-    name: "ariso-stt",
+    name: "ariso-stt-mac",
     platforms: [.macOS(.v14)],
     dependencies: [
         .package(url: "https://github.com/FluidInference/FluidAudio.git", from: "0.14.8"),
