+++ src-tauri/ariso-stt-mac/Sources/ariso-stt/main.swift
