+++ src-tauri/ariso-stt-shared/fixtures/transcript.json
@@ -1,0 +1,18 @@
+{
+  "language": "en",
+  "durationSeconds": 12.0,
+  "participants": [
+    {
+      "id": 0,
+      "label": "Speaker 1"
+    }
+  ],
+  "segments": [
+    {
+      "speaker": 0,
+      "text": "Hello from oats.",
+      "start": 0.0,
+      "end": 1.5
+    }
+  ]
+}
