+++ src-tauri/ariso-stt-windows/Cargo.toml
@@ -1,19 +1,16 @@
 [package]
-name = "ariso-stt"
+name = "ariso-stt-windows"
 version = "0.10.0"
 edition = "2024"
 
+[[bin]]
+name = "ariso-stt"
+path = "src/main.rs"
+
 [dependencies]
 anyhow = "1"
-bzip2 = "0.4"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-sha2 = "0.10"
 sherpa-onnx = "1.13.3"
 symphonia = { version = "0.5", default-features = false, features = ["mp3", "wav", "pcm"] }
-tar = "0.4"
-ureq = "2"
-zip = "4"
-
-[dev-dependencies]
 tempfile = "3"
