--- thoughts/shared/prs/173_description.md
@@ -1,46 +1,0 @@
-## Summary
-- replace the Windows `ariso-stt` placeholder with a Windows-native Local spike built around Parakeet-family STT, sherpa-onnx diarization, and Gemma-family notes through llama.cpp
-- enable Windows Local capability/model readiness in the product seams so Settings can select Local, show installed model rows, and reuse the existing retry/finalize commands
-- add Windows model validation/download hooks, versioned `models/windows/.../v1` layout support, and scripts for syncing/smoke-testing the Windows model bundle
-- update desktop/release CI inputs so the real Windows sidecar is built, copied into Tauri's `externalBin` location, and packaged with the app
-- add an internal MSI artifact path alongside the existing NSIS artifact while keeping NSIS as the primary consumer installer
-- document the Parakeet Local parity spike, Windows model layout, verification flow, and remaining release-hardening boundaries
-
-## What changed
-- `src-tauri/ariso-stt-cross` now implements:
-  - `ariso-stt --audio <path> --models <dir> --format json`
-  - `ariso-stt download --models <dir>`
-  - `ariso-stt download-notes --models <dir>`
-  - `ariso-stt notes --transcript <path> --models <dir>`
-- Transcription emits the existing JSON contract: `language`, `participants`, and speaker-attributed `segments`.
-- Notes emit Markdown on stdout and fail on empty output.
-- Windows model readiness recognizes the Windows Parakeet, diarization, and Gemma/llama.cpp artifact layout.
-- Windows `PlatformCapabilities.localBackend` is now supported with `cpp-sidecar`.
-- Added `scripts/windows-local-smoke.ps1` and `scripts/sync-windows-local-models.ps1`.
-- Added `scripts/build-windows-sidecar.ps1` so desktop/release CI share the same Windows sidecar copy step.
-- Added `src-tauri/tauri.windows.conf.json` so Windows no longer needs a fake `mlx-swift_Cmlx.bundle` resource placeholder.
-- Release CI now builds `nsis,msi`, verifies both installer outputs and updater signatures, and uploads `windows-nsis-internal` plus `windows-msi-internal`.
-- Added the Windows schema generated by the Tauri capability/schema updates.
-
-## Verification
-- [x] `cargo +stable-x86_64-pc-windows-msvc test --manifest-path src-tauri/ariso-stt-cross/Cargo.toml --locked`
-- [x] `git diff --check`
-- [x] Windows sidecar CLI smoke: Parakeet transcription against the live smoke audio/model root
-- [x] Windows sidecar notes smoke: Gemma/llama.cpp generated Markdown notes through `scripts/windows-local-smoke.ps1`
-- [x] Windows sidecar helper smoke: `powershell -ExecutionPolicy Bypass -File scripts\build-windows-sidecar.ps1`
-- [x] Windows installer bundle smoke: `powershell -ExecutionPolicy Bypass -File scripts\build-windows-installers.ps1`
-- [x] Windows installer artifact check: NSIS `.exe`, MSI `.msi`, and updater `.sig` files exist under `src-tauri/target/x86_64-pc-windows-msvc/release/bundle`
-- [x] Live app smoke: Settings showed `Local` selected with Speech and Language model rows `Installed`
-- [x] Live app smoke: the app invoked `retry_local_transcription`, spawned `ariso-stt.exe`, wrote `transcript.md`, updated `meta.json` to `done`, and generated `ari-note.md`
-- [x] Release launch smoke: target release and QA-installed `oats.exe` stayed alive after launch
-
-## Smoke evidence
-- Live transcript artifact: `.codex-artifacts/live-smoke-root/recordings/2026-07-10T12-00-00Z/transcript.md`
-- Live notes artifact: `.codex-artifacts/live-smoke-root/recordings/2026-07-10T12-00-00Z/ari-note.md`
-- The live retry meta recorded `language: en`, two participants, and `modelVersion: parakeet-tdt-0.6b-v3`.
-
-## Known boundaries
-- This remains a Windows Local spike, not public Windows release readiness.
-- MSI remains internal until signing, updater, install, upgrade, and uninstall behavior are smoke tested.
-- Local installer smoke used a throwaway updater key, so the expected key-mismatch warning is not release evidence.
-- Windows system-audio loopback and Windows auto-record are outside this spike's plan scope.
